requirements.txt
langchain
python-dotenv
langchain_community
langchain-text-splitters
sentence-transformers
pypdf
Faiss
langchain_huggingface
